backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets>=13.0
openai>=1.55.0
supabase>=2.0.0
python-dotenv==1.0.0
numpy<2.0